Add clang-tidy configuration and fix key warnings
CLANG-TIDY SETUP:
- Updated .clang-tidy with VM-focused rules emphasizing performance and safety
- Enabled bugprone, cert, performance, and readability checks
- Disabled overly pedantic rules inappropriate for VM code
- Configured check options for performance-critical code

KEY FIXES APPLIED:
- Fix memory manager multiplication to use explicit size_t casting
- Update PrimitiveFailure constructor to use move semantics
- Fix TaggedValue constructor to use member initializer list
- Update Symbol constructor to use move semantics
- Fix implicit bool conversions in pointer checks
- Add explicit nullptr comparisons for better readability
- Initialize variables to prevent undefined behavior
- Pre-allocate vector capacity for performance

CLANG-TIDY RESULTS:
- Array primitives: Clean (only reference member warning from header)
- Object primitives: Clean (only reference member warning from header)
- Significant reduction in warnings across codebase
- All tests still pass, no functionality affected

This establishes better code quality standards for the VM while maintaining performance.

+++ src/cpp/include/memory_manager.h
@@ -27,7 +27,7 @@ public:
      * 
      * @param initialSpaceSize The size of each memory space in bytes.
      */
-    static constexpr size_t DEFAULT_INITIAL_SPACE_SIZE = 1024 * 1024; // 1MB
+    static constexpr size_t DEFAULT_INITIAL_SPACE_SIZE = static_cast<size_t>(1024) * 1024; // 1MB
 
     /**
      * Constructor that initializes memory spaces.
